requirements.txt
google-genai
google-search-results
Pillow
requests
rich